services/worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic==2.8.2
pydantic-settings==2.4.0

# Google Drive API
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0

# Supabase
supabase==2.6.0

# PDF parsing (after OCR)
pypdf==5.0.1
